fix(#170): check version when refreshing cache
